requirements.txt
torch>=2.1.0
transformers>=4.35.2
google-api-python-client>=2.108.0
python-dotenv>=1.0.0
numpy>=1.24.0
tqdm>=4.66.1
pydantic>=2.5.1
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6 